Added support for some kinds of text alignment-baseline.
This patch provides a partial fix for the missing formatting
of the "alignment-baseline" attribute for the "text" SVG
tag.  It adds most of the alignment baselines, although it
omits the language and text-specific ones; indeed, only
Western text will be aligned. The "middle" baseline is
treated as "central" due to the lack of access to font
x-height information in Cairo.