Added non-standard "display-anchor" tag to align text vertically.
The "display-anchor" attribute of the "text" tag is an
invention of the patch author. It provides the ability to
align text (as opposed to just a font) vertically at either
the "top", "middle", or "bottom" of the text. If
display-anchor and alignment-baseline are both provided,
display-anchor takes precedence. The display-anchor
attribute is only processed for Western fonts.