Add support to PDFs with multiple pages
I kept the interface the same, for compatibility with existing code, but adding the option to pass an array of bytestring, file_obj, url, or tree objects (this last one as an optimization when converting the same svg to multiple formats), instead of a single object only. 
For PDFs it should call the method addPage for each new page, except the first one. For all other formats it should raise NotImplementedError for now.