Don’t use overlapping groups for regular expressions
The section between 'rgb(' and the final ')' contains multiple overlapping
groups.

Since all three infinitely repeating groups accept spaces, a long string of
spaces causes catastrophic backtracking when it is not followed by a closing
parenthesis.

The complexity is cubic, so doubling the length of the malicious string of
spaces makes processing take 8 times as long.